# src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: 온라인 도서 관리 시스템 API
  description: 도서 등록, 조회, 검색, 태그 필터링, 대출 관리 REST API
  version: 1.0.0

paths:
  /api/books:
    get:
      summary: 모든 도서 조회
      responses:
        200:
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
    post:
      summary: 새로운 도서 등록
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
      responses:
        201:
          description: 도서 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"

  /api/books/{id}:
    get:
      summary: 특정 도서 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
    put:
      summary: 특정 도서 수정
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
      responses:
        200:
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
    delete:
      summary: 특정 도서 삭제
      responses:
        204:
          description: 삭제 성공

  /api/books/search:
    get:
      summary: 도서 검색 (제목 또는 저자)
      parameters:
        - name: type
          in: query
          required: false
          description: 검색 유형 (title 또는 author), 기본값은 title
          schema:
            type: string
            enum: [title, author]
        - name: keyword
          in: query
          required: true
          description: 검색 키워드
          schema:
            type: string
      responses:
        200:
          description: 검색 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"

  /api/books/filter:
    get:
      summary: 태그 기반 도서 필터링
      parameters:
        - name: tags
          in: query
          required: true
          description: 필터링할 태그 목록 (여러 개 가능)
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: 태그 필터링 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"

  /api/users:
    get:
      summary: 모든 사용자 조회
      responses:
        200:
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      summary: 새로운 사용자 등록
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        201:
          description: 사용자 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /api/users/{id}:
    get:
      summary: 특정 사용자 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /api/loans:
    post:
      summary: 도서 대출
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanRequest"
      responses:
        201:
          description: 대출 성공

  /api/loans/{bookId}:
    get:
      summary: 특정 도서의 대출 상태 확인
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 대출 상태 확인 성공

  /api/loans/{bookId}/return:
    post:
      summary: 도서 반납
      responses:
        200:
          description: 반납 성공

components:
  schemas:
    BookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        tags:
          type: array
          items:
            type: string

    BookResponse:
      type: object
      properties:
        id:
          type: integer
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        available:
          type: boolean
        tags:
          type: array
          items:
            type: string

    UserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string

    LoanRequest:
      type: object
      properties:
        userId:
          type: integer
        bookId:
          type: integer